+++ FretFrameworkV2/Song Library/Entry/Extended_Scan_Drums.h
@@ -9,7 +9,7 @@ template<>
 bool Extended_Scan::DifficultyTracker<DrumNote<DrumPad, 5>>::testSingleNote(ScanValues& values, size_t lane) noexcept;
 
 template<>
-bool Extended_Scan::DifficultyTracker<DrumNote<DrumPad_Pro, 4>>::testMultiNote(ScanValues& values, const std::vector<std::pair<size_t, uint32_t>>& colors) noexcept;
+bool Extended_Scan::DifficultyTracker<DrumNote<DrumPad_Pro, 4>>::testMultiNote(ScanValues& values, const std::vector<std::pair<size_t, uint64_t>>& colors) noexcept;
 
 template<>
-bool Extended_Scan::DifficultyTracker<DrumNote<DrumPad, 5>>::testMultiNote(ScanValues& values, const std::vector<std::pair<size_t, uint32_t>>& colors) noexcept;
+bool Extended_Scan::DifficultyTracker<DrumNote<DrumPad, 5>>::testMultiNote(ScanValues& values, const std::vector<std::pair<size_t, uint64_t>>& colors) noexcept;
